feat(test): isolate test collection to file
