+++ packages/firefuel/test/src/utils/query_extensions_test.dart
@@ -2,8 +2,8 @@ import 'package:fake_cloud_firestore/fake_cloud_firestore.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:firefuel/firefuel.dart';
+import '../../utils/test_collection.dart';
 import '../../utils/test_user.dart';
-import '../firefuel_collection_test.dart';
 
 void main() {
   late TestCollection testCollection;
